(Refactor) Move more types to types.dart
These types are referenced in more than one file.
