+++ lib/src/chatoptions.dart
@@ -1,131 +1,6 @@
 import 'dart:convert';
 
-import 'package:talkjs_flutter/src/themeoptions.dart';
-
-import './chatbox.dart';
-
-/// The values that dictate the chat direction.
-enum TextDirection {
-  /// right-to-left
-  rtl,
-
-  /// left-to-right
-  ltr,
-}
-
-/// Settings that affect the behavior of the message field
-class MessageFieldOptions {
-  /// Determines whether the message field should automatically focus when the
-  /// user navigates.
-  ///
-  /// Defaults to "smart", which means that the message field gets focused
-  /// whenever a conversation is selected, if possible without negative side
-  /// effects.
-  /// If you need more control, consider setting [autofocus] to false and
-  /// calling focus() at appropriate times.
-  final bool? autofocus; // Convert to "smart"
-
-  /// If set to true, pressing the enter key sends the message
-  /// (if there is text in the message field).
-  ///
-  /// When set to false, the only way to send a message is by clicking or
-  /// touching the "Send" button.
-  /// Defaults to true.
-  final bool? enterSendsMessage;
-
-  /// The text displayed in the message field when the user hasn't started
-  /// typing anything.
-  final String? placeholder;
-
-  /// This enables spell checking.
-  ///
-  /// Note that setting this to true may also enable autocorrect on some mobile
-  /// devices.
-  /// Defaults to false
-  final bool? spellcheck;
-
-  /// TODO: visible
-
-  const MessageFieldOptions({
-    this.autofocus,
-    this.enterSendsMessage,
-    this.placeholder,
-    this.spellcheck,
-  });
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> result = {
-      'enterSendsMessage': ?enterSendsMessage,
-      'placeholder': ?placeholder,
-      'spellcheck': ?spellcheck,
-    };
-
-    if (autofocus != null) {
-      if (autofocus == true) {
-        result['autofocus'] = 'smart';
-      } else {
-        result['autofocus'] = autofocus;
-      }
-    }
-
-    return result;
-  }
-
-  bool operator ==(Object other) {
-    if (identical(this, other)) {
-      return true;
-    }
-
-    if (other is! MessageFieldOptions) {
-      return false;
-    }
-
-    if (autofocus != other.autofocus) {
-      return false;
-    }
-
-    if (enterSendsMessage != other.enterSendsMessage) {
-      return false;
-    }
-
-    if (placeholder != other.placeholder) {
-      return false;
-    }
-
-    if (spellcheck != other.spellcheck) {
-      return false;
-    }
-
-    return true;
-  }
-
-  int get hashCode =>
-      Object.hash(autofocus, enterSendsMessage, placeholder, spellcheck);
-}
-
-/// The possible values for showTranslationToggle
-enum TranslationToggle { off, on, auto }
-
-extension TranslationToggleValue on TranslationToggle {
-  /// Converts this enum's values to String.
-  dynamic getValue() => switch (this) {
-    TranslationToggle.off => false,
-    TranslationToggle.on => true,
-    TranslationToggle.auto => 'auto',
-  };
-}
-
-/// The possible values for translateConversations
-enum TranslateConversations { off, on, auto }
-
-extension TranslateConversationsValue on TranslateConversations {
-  /// Converts this enum's values to String.
-  dynamic getValue() => switch (this) {
-    TranslateConversations.off => false,
-    TranslateConversations.on => true,
-    TranslateConversations.auto => 'auto',
-  };
-}
+import './types.dart';
 
 /// Options to configure the behaviour of the [ChatBox] UI.
 class ChatBoxOptions {
