--- lib/src/themeoptions.dart
@@ -1,8 +1,0 @@
-class ThemeOptions {
-  final String? name;
-  final Map<String, dynamic>? custom;
-
-  const ThemeOptions({this.name, this.custom});
-
-  Map<String, dynamic> toJson() => {'name': ?name, 'custom': ?custom};
-}
