+++ lib/talkjs_flutter.dart
@@ -13,7 +13,6 @@ export 'src/user.dart';
 export 'src/conversationlist.dart';
 export 'src/predicate.dart';
 export 'src/notification.dart';
-export 'src/themeoptions.dart';
 export 'src/unreads.dart';
 export 'src/types.dart';
